// frontend/src/app/core/api/model/securityAlertDTO.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SecurityAlertDTO { 
    id?: number;
    alertType?: string;
    title?: string;
    message?: string;
    severity?: string;
    createdAt?: string;
    read?: boolean;
}